ovi/backend/requirements.txt
Flask  
datetime 
uuid
PyJWT==1.7.1